+++ frontend/src/components/cards/GraphCard.vue
@@ -104,3 +104,9 @@ const chartProps = computed(() => {
     <LineChart v-bind="chartProps" />
   </BaseCard>
 </template>
+
+<style scoped>
+:deep(.card-content) {
+  height: 100%;
+}
+</style>